Give ability for users to get uncertainty values.
