+++ experiments/nlp_bert_mcdropout.py
@@ -15,7 +15,8 @@ from datasets import load_dataset
 from transformers import BertTokenizer, TrainingArguments
 from transformers import BertForSequenceClassification
 
-from baal.active import get_heuristic, active_huggingface_dataset, HuggingFaceDatasets
+from baal.active import get_heuristic
+from baal.active.nlp_datasets import active_huggingface_dataset, HuggingFaceDatasets
 from baal.active.active_loop import ActiveLearningLoop
 from baal.bayesian.dropout import patch_module
 from baal.transformers_trainer_wrapper import BaalTransformersTrainer
