+++ src/baal/active/__init__.py
@@ -1,10 +1,9 @@
-from typing import Union, Callable, Dict
+from typing import Union, Callable
 
 from . import heuristics
 from .active_loop import ActiveLearningLoop
 from .dataset import ActiveLearningDataset
 from .file_dataset import FileDataset
-from .nlp_datasets import HuggingFaceDatasets
 
 
 def get_heuristic(name: str, shuffle_prop: float = 0.0,
@@ -34,29 +33,3 @@ def get_heuristic(name: str, shuffle_prop: float = 0.0,
     }[name](shuffle_prop=shuffle_prop, reduction=reduction, **kwargs)
 
 
-def active_huggingface_dataset(dataset,
-                               tokenizer=None,
-                               target_key: str = "label",
-                               input_key: str = "sentence",
-                               max_seq_len: int = 128,
-                               **kwargs):
-    """
-    Wrapping huggingface.datasets with baal.active.ActiveLearningDataset.
-
-    Args:
-        dataset (torch.utils.data.Dataset): a dataset provided by huggingface.
-        tokenizer (transformers.PreTrainedTokenizer): a tokenizer provided by huggingface.
-        target_key (str): target key used in the dataset's dictionary.
-        input_key (str): input key used in the dataset's dictionary.
-        max_seq_len (int): max length of a sequence to be used for padding the shorter sequences.
-        kwargs (Dict): Parameters forwarded to 'ActiveLearningDataset'.
-
-    Returns:
-        an baal.active.ActiveLearningDataset object.
-    """
-
-    return ActiveLearningDataset(HuggingFaceDatasets(dataset,
-                                                     tokenizer,
-                                                     target_key,
-                                                     input_key,
-                                                     max_seq_len), **kwargs)
